making modifications to the OtherClient and OtherServer. Renamed to Client and Server
