+++ blockchain/src/Server.java
@@ -1,56 +1,143 @@
-import java.io.*;
-import java.net.*;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.PrintWriter;
+import java.net.ServerSocket;
+import java.net.Socket;
+
+/**
+ * A server program which accepts requests from clients to
+ * capitalize strings.  When clients connect, a new thread is
+ * started to handle an interactive dialog in which the client
+ * sends in a string and the server thread sends back the
+ * capitalized version of the string.
+ *
+ * The program is runs in an infinite loop, so shutdown in platform
+ * dependent.  If you ran it from a console window with the "java"
+ * interpreter, Ctrl+C generally will shut it down.
+ */
 public class Server implements Runnable{
-    int server_Port = 5555;
-    String server_IP;
 
-    static Server server;
-    private Server(){
-        try {
-            InetAddress iAddress = InetAddress.getLocalHost();
-            server_IP = iAddress.getHostAddress();
-            System.out.println("Server IP address : " + server_IP);
-        } catch (IOException e) {
+    /**
+     * Application method to run the server runs in an infinite loop
+     * listening on port 9898.  When a connection is requested, it
+     * spawns a new thread to do the servicing and immediately returns
+     * to listening.  The server keeps a unique client number for each
+     * client that connects just to show interesting logging
+     * messages.  It is certainly not necessary to do this.
+     */
 
+    private static Server server;
+    private ServerSocket listener;
+
+    private Server() throws Exception{
+        System.out.println("The capitalization server is running.");
+        ServerSocket listener = new ServerSocket(9898);
+    }
+
+    public static Server getServer(){
+        if(server == null){
+            try{
+                server = new Server();
+            }catch (Exception e){
+                return null;
+            }
         }
+        return server;
+    }
 
-   }
-
-
-   public static Server getServer(){
-       if(server == null){
-           server = new Server();
-       }
-       return server;
-   }
+    /**
+     * When an object implementing interface <code>Runnable</code> is used
+     * to create a thread, starting the thread causes the object's
+     * <code>run</code> method to be called in that separately executing
+     * thread.
+     * <p>
+     * The general contract of the method <code>run</code> is that it may
+     * take any action whatsoever.
+     *
+     * @see Thread#run()
+     */
     @Override
     public void run() {
-            try {
-                ServerSocket serverSocket = new ServerSocket(server_Port);
+        int clientNumber = 0;
 
-                while (true) {
+        try {
+            while (true) {
+                new Capitalizer(listener.accept(), clientNumber++).start();
+            }
+        }catch (Exception e) {
 
-                    Socket socket = serverSocket.accept();
+        }finally
+         {
+             try{
+                 listener.close();
+             }catch (Exception e){
 
-                    OutputStream os = socket.getOutputStream();
-                    PrintWriter pw = new PrintWriter(os, true);
-                    InputStreamReader isr = new InputStreamReader(socket.getInputStream());
-
-                    pw.println("Connection confirmed ");
-
-                    BufferedReader br = new BufferedReader(isr);
-                    String str = br.readLine();
-
-                    pw.println("your ip address is " + str);
-
-                    pw.close();
-                }
-            }catch (Exception e){
-                System.out.println(e);
-
+             }
         }
     }
 
+    /**
+     * A private thread to handle capitalization requests on a particular
+     * socket.  The client terminates the dialogue by sending a single line
+     * containing only a period.
+     */
+    private static class Capitalizer extends Thread {
+        private Socket socket;
+        private int clientNumber;
 
+        public Capitalizer(Socket socket, int clientNumber) {
+            this.socket = socket;
+            this.clientNumber = clientNumber;
+            log("New connection with client# " + clientNumber + " at " + socket);
+        }
 
+        /**
+         * Services this thread's client by first sending the
+         * client a welcome message then repeatedly reading strings
+         * and sending back the capitalized version of the string.
+         */
+        public void run() {
+            try {
+
+                // Decorate the streams so we can send characters
+                // and not just bytes.  Ensure output is flushed
+                // after every newline.
+                BufferedReader in = new BufferedReader(
+                        new InputStreamReader(socket.getInputStream()));
+                PrintWriter out = new PrintWriter(socket.getOutputStream(), true);
+
+                // Send a welcome message to the client.
+                out.println("Hello, you are client #" + clientNumber + ".");
+                out.println("Enter a line with only a period to quit\n");
+
+                // Get messages from the client, line by line; return them
+                // capitalized
+                while (true) {
+                    String input = in.readLine();
+                    if (input == null || input.equals(".")) {
+                        break;
+                    }
+                    out.println(input.toUpperCase());
+                }
+            } catch (IOException e) {
+                log("Error handling client# " + clientNumber + ": " + e);
+            } finally {
+                try {
+                    socket.close();
+                } catch (IOException e) {
+                    log("Couldn't close a socket, what's going on?");
+                }
+                log("Connection with client# " + clientNumber + " closed");
+            }
+        }
+
+        /**
+         * Logs a simple message.  In this case we just write the
+         * message to the server applications standard output.
+         */
+        private void log(String message) {
+            System.out.println(message);
+        }
+    }
 }
